+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 * jdbc: set UpdatetableEntity.updatedAt to millisecond precision to avoid issues with rapid re-saving of entities
 * json: change detection on how to render inline classes based on String constructor and cache the knowledge in JsonMapper instance for better performance
+* server: add backwards-compatibility to KeyCipher, so that CookieSessionStore can still decrypt old cookies encrypted with ECB mode, but new cookies will be encrypted with GCM mode for better security introduced in 1.8.2
 
 # 1.8.2
 * JDK 17+ is now required